requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
future==0.17.1
gTTS==2.0.3
gTTS-token==1.1.3
idna==2.8
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
pydub==0.23.1
Pygments==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pyzmq==18.1.0
qtconsole==4.5.4
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.3
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
